Unify duplicated render/stats flow between main.rs and lib.rs (#32)
main() reimplemented the ~155-line stats-update flow (session cost update →
max-token tracking → adaptive context learning) instead of sharing it with the
public render_statusline() API. The two copies had already drifted (device_id
handling; max-token tracking nested inside the cost branch in main but top-level
in lib), so fixes to one path could silently miss the other.

Extract the flow into a single shared module, src/render.rs
(update_stats_and_daily_total), called by both main()'s default path and
render_statusline(). The binary and library are separate crates that each
compile the module files, so the helper lives in a shared file declared in both
crate roots rather than in a crate root itself.

- Removes the session_id.unwrap() in the render path (destructured guard instead)
- Aligns main to lib's cost-independent max-token tracking (more correct for
  compaction detection; no-op in practice since cost is always present)
- Keeps each path's own output call: main uses format_output (layout/template
  aware), render_statusline uses format_output_to_string (intentionally not)

Net ~270 duplicated lines collapsed into one ~149-line module.

Verified: full suite 1059 passed / 0 failed / 19 ignored; clippy -D warnings
clean; cargo fmt --check clean; binary smoke test renders correctly.
