+++ src/lib.rs
@@ -54,6 +54,8 @@ pub mod migrations;
 pub mod models;
 /// Data provider system for parallel variable collection
 pub mod provider;
+/// Shared statusline rendering logic (stats-update flow used by the binary and the embedding API)
+pub mod render;
 /// Retry logic with exponential backoff for transient failures
 pub mod retry;
 /// Hook-based state management for real-time event tracking
@@ -138,133 +140,8 @@ pub fn render_statusline(input: &StatuslineInput, update_stats: bool) -> Result<
     // Get session ID
     let session_id = input.session_id.as_deref();
 
-    // Load or update stats
-    let daily_total = if update_stats && session_id.is_some() {
-        // Update stats with new data
-        if let Some(ref cost) = input.cost {
-            if let Some(total_cost) = cost.total_cost_usd {
-                // Extract model name and workspace directory
-                let model_name = input
-                    .model
-                    .as_ref()
-                    .and_then(|m| m.display_name.as_ref())
-                    .map(|s| s.as_str());
-                let workspace_dir = input
-                    .workspace
-                    .as_ref()
-                    .and_then(|w| w.current_dir.as_ref())
-                    .map(|s| s.as_str());
-
-                // Extract token breakdown from transcript if available
-                let token_breakdown = input
-                    .transcript
-                    .as_ref()
-                    .and_then(|path| utils::get_token_breakdown_from_transcript(path));
-
-                // Get device ID for audit trail
-                let device_id = common::get_device_id();
-
-                use crate::database::SessionUpdate;
-                let (daily_total, _monthly_total) = stats::update_stats_data(|data| {
-                    data.update_session(
-                        session_id.unwrap(),
-                        SessionUpdate {
-                            cost: total_cost,
-                            lines_added: cost.total_lines_added.unwrap_or(0),
-                            lines_removed: cost.total_lines_removed.unwrap_or(0),
-                            model_name: model_name.map(|s| s.to_string()),
-                            workspace_dir: workspace_dir.map(|s| s.to_string()),
-                            device_id: Some(device_id),
-                            token_breakdown,
-                            max_tokens_observed: None, // updated separately
-                            // active_time_seconds / last_activity are owned and computed by
-                            // SqliteDatabase::update_session (src/database/session.rs); not available at
-                            // this SessionUpdate construction site. v3.0.0 cleanup per CONTEXT.md D-13.
-                            active_time_seconds: None,
-                            last_activity: None,
-                        },
-                    )
-                });
-                daily_total
-            } else {
-                // Have session but no cost data - still load existing daily totals
-                let data = stats::get_or_load_stats_data();
-                stats::get_daily_total(&data)
-            }
-        } else {
-            // No cost data - just get current daily total
-            let data = stats::get_or_load_stats_data();
-            stats::get_daily_total(&data)
-        }
-    } else {
-        // Just get current daily total without updating
-        let stats_data = stats::get_or_load_stats_data();
-        stats::get_daily_total(&stats_data)
-    };
-
-    // Track max_tokens_observed for compaction detection
-    // This runs regardless of adaptive_learning setting
-    if update_stats {
-        if let (Some(transcript), Some(session)) = (transcript_path, session_id) {
-            if let Some(current_tokens) = utils::get_token_count_from_transcript(transcript) {
-                // Update session's max_tokens_observed
-                // This updates both in-memory stats and SQLite database
-                stats::update_stats_data(|data| {
-                    data.update_max_tokens(session, current_tokens);
-                    // Return unchanged totals (we're just updating token tracking)
-                    use crate::common::{current_date, current_month};
-                    let today = current_date();
-                    let month = current_month();
-                    let daily_total = data.daily.get(&today).map(|d| d.total_cost).unwrap_or(0.0);
-                    let monthly_total = data
-                        .monthly
-                        .get(&month)
-                        .map(|m| m.total_cost)
-                        .unwrap_or(0.0);
-                    (daily_total, monthly_total)
-                });
-
-                // Adaptive context learning: observe token usage if enabled
-                if let Some(model) = model_name {
-                    let config = config::get_config();
-                    if config.context.adaptive_learning {
-                        // Get previous token count from session stats
-                        let stats_data = stats::get_or_load_stats_data();
-                        let previous_tokens = stats_data
-                            .sessions
-                            .get(session)
-                            .and_then(|s| s.max_tokens_observed)
-                            .map(|t| t as usize);
-
-                        // Create context learner and observe usage
-                        use crate::common::get_data_dir;
-                        use crate::context_learning::ContextLearner;
-                        use crate::database::SqliteDatabase;
-
-                        let db_path = get_data_dir().join("stats.db");
-                        if let Ok(db) = SqliteDatabase::new(&db_path) {
-                            let learner = ContextLearner::new(db);
-                            // Extract workspace_dir and device_id for audit trail
-                            let workspace_dir = input
-                                .workspace
-                                .as_ref()
-                                .and_then(|w| w.current_dir.as_deref());
-                            let device_id = crate::common::get_device_id();
-                            // Ignore errors from adaptive learning - it's experimental and shouldn't block statusline
-                            let _ = learner.observe_usage(
-                                model,
-                                current_tokens as usize,
-                                previous_tokens,
-                                Some(transcript),
-                                workspace_dir,
-                                Some(&device_id),
-                            );
-                        }
-                    }
-                }
-            }
-        }
-    }
+    // Load or update stats (single shared implementation; see src/render.rs).
+    let daily_total = render::update_stats_and_daily_total(input, update_stats);
 
     // Format the output to string
     let output = display::format_output_to_string(
